Extract LinkType to abi/wasm-object.h
